graph auf double geupdatet, addpoint funktion

+++ src/view/Punkt.java
@@ -2,18 +2,18 @@ package view;
 
 public class Punkt 
 {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
-	public int getY() 
+	public double getY() 
 	{
 		return y;
 	}
 
-	public Punkt(int x, int y)
+	public Punkt(double x, double y)
 	{
 		this.x = x;
 		this.y = y;
